WIP: add buildmode
